Fixed bug; addressed warnings; updated parse-trace
- Bug fix

  - accessed DataMsg::field_array_list_ values when vector size 0
    (optimized out for O3 so only shows up with <= O2)

- Warnings

  - out-of-order constructor initializations
  - comparing signed/unsigned
  - removed unaccessed variables
  - removed unaccessed methods
  - possibly uninitialized values
  - printf formats
  - missing return
  - avoid accessing Memory::instance() when CONFIG_USE_MEMORY
    not defined

- EZTrace

  - fixed parse-trace for multiple cycles

- MethodOrderHilbert

  - updated index/count to long long to match MethodOrder

+++ src/Cello/control_adapt.cpp
@@ -262,7 +262,7 @@ int Block::adapt_compute_desired_level_(int level_maximum)
   int level = this->level();
   int level_desired = level;
 
-  for (int k=0; k<cello::num_refine(); k++) {
+  for (size_t k=0; k<cello::num_refine(); k++) {
 
     Schedule * schedule = cello::refine(k)->schedule();
